+++ upydev/commandlib.py
@@ -113,6 +113,7 @@ SET_LOCALNAME = "f=open('localname.py','wb');f.write(b'{}');f.close();"
 SHASUM_CHECK = "from shasum import shasum_check;shasum_check('{}');gc.collect()"
 SHASUM = "from shasum import shasum;shasum('{}');gc.collect()"
 LS = "import os;os.listdir('{}');gc.collect()"
+CAT = "cat('{}');gc.collect()"
 
 CMDDICT_ = {'UID': UID, 'UPYSH': UPYSH, 'HELP': HELP, 'MOD': MODULES,
             'MEM': MEM, 'OS_STAT': OS_STAT, 'FILE_STAT': FILE_STAT,
@@ -132,6 +133,6 @@ CMDDICT_ = {'UID': UID, 'UPYSH': UPYSH, 'HELP': HELP, 'MOD': MODULES,
             'SD_DEINIT': SD_DEINIT, 'SD_AUTO': SD_AUTO, 'CHECK_UPYSH2': CHECK_UPYSH2,
             'SET_HOSTNAME': SET_HOSTNAME, 'SET_LOCALNAME': SET_LOCALNAME,
             'SHASUM_CHECK': SHASUM_CHECK,
-            'SHASUM': SHASUM, 'LS': LS}
+            'SHASUM': SHASUM, 'LS': LS, 'CAT': CAT}
 
 _CMDDICT_ = {k: 'import gc;' + v for k, v in CMDDICT_.items()}